requirements.txt
attrs==21.4.0
atomicwrites==1.4.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.3
pip==21.1.2
pluggy==1.0.0
pyparsing==3.0.7
py==1.11.0
pytest==7.1.0
setuptools==57.0.0
tomli==2.0.1
wheel==0.36.2
requests==2.27.1
selenium~=4.0.0
Appium-Python-Client~=2.1.4